Add common include file that now incldues file used for syntax highlighting. Set common output folder and intermediate files for all projects Fix some warnings when building code

+++ 01/08-grid-stride-double/03-grid-stride-double.cu
@@ -1,4 +1,6 @@
 #include <stdio.h>
+#include "common.h"
+
 
 void init(int *a, int N)
 {
@@ -54,8 +56,8 @@ int main()
    * The size of this grid is 256*32 = 8192.
    */
 
-  size_t threads_per_block = 256;
-  size_t number_of_blocks = 32;
+  int threads_per_block = 256;
+  int number_of_blocks = 32;
 
   doubleElements << <number_of_blocks, threads_per_block >> > (a, N);
   cudaDeviceSynchronize();
